[ARC-535] Organize custom resources module
around domain instead of by kind

branch: feature/ARC-535-target-specific-dictionary
